+++ lib/calculator.dart
@@ -1,0 +1,17 @@
+library flutter_image_picker;
+
+/// A Calculator.
+class Calculator {
+  /// Returns [value] plus 1.
+  int addOne(int value) => value + 1;
+
+  int substractOne(int value) => value - 1;
+
+  int valueTimesTen(int value) {
+    return value * 10;
+  }
+
+  int valueDividedByTen(int value) {
+    return (value / 10).round();
+  }
+}