+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 * Added more folders
+
+# 0.0.5
+
+* Added image picker package which shows a dialog to choose an image from gallery or to make a picture with the camera
+* Added example app to test/show off the image picker package functionality